There is a lot in this merge. Mainly adding the Game_systems table and linking it by ID to the master data structures.
That means changing tests.
During the process some fixes aside, not related to the topic came in.

+++ backend/importer/exporter_test.go
@@ -25,7 +25,7 @@ func TestExportChar2VTT(t *testing.T) {
 	assert.NoError(t, err, "Expected no error when exporting char")
 
 	// Basic validations
-	assert.Equal(t, char.UserID, 1)
+	assert.Equal(t, char.UserID, uint(1))
 	assert.Equal(t, char.Name, exportedChar.Name)
 	assert.Equal(t, char.Rasse, exportedChar.Rasse)
 	assert.Equal(t, char.Typ, exportedChar.Typ)
